pyproject.toml
[tool.poetry]
name = "raceid-importer"
version = "0.1.0"
description = ""
authors = ["Sebastian Ullrich <sebastian.ullrich@deepsource.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
garminconnect = "^0.2.30"
termcolor = "^3.2.0"
rich = "^14.2.0"
textual = "^6.4.0"
pyperclip = "^1.11.0"
python-dateutil = "^2.9.0.post0"
python-debouncer = "^0.1.5"
requests = "^2.32.5"
textual-dev = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
